Refactor: Move page widgets to widgets directory
Moves home_page.dart, manage_models_page.dart, and settings_page.dart
from lib/ to lib/widgets/ to better organize UI components.
Updates all relevant imports.
